refactor: OrderGroup 리팩터링
